Ajustes na versão 1

+++ api/src/application/usecases/users/delete-user.usecase.ts
@@ -1,9 +1,10 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { Inject, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import type { UserRepository } from '../../../domain/repositories/user.repository';
 import { USER_REPOSITORY } from 'src/application/injection-tokens/user-repository.token';
 
 interface DeleteUserInput {
     id: number;
+    userId: number
 }
 
 @Injectable()
@@ -13,13 +14,14 @@ export class DeleteUserUseCase {
         private readonly userRepository: UserRepository,
     ) { }
 
-    async execute(input: DeleteUserInput): Promise<void> {
-        const exists = await this.userRepository.findById(input.id);
-
-        if (!exists) {
-            throw new Error('User not found');
+    async execute(input: DeleteUserInput ): Promise<void> {
+        if( input.id !== input.userId ) {
+            throw new UnauthorizedException( "Unauthorized" );
         }
-
+        const exists = await this.userRepository.findById(input.id);
+        if (!exists) {
+            throw new NotFoundException('User not found');
+        }
         await this.userRepository.delete(input.id);
     }
 }
